customers/serializers.py
from rest_framework import serializers
from customers.models import Customer


class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        exclude = ("updated_by", "created_by", "created_at", "updated_at")
