requirements.txt
django
djangorestframework
celery
psycopg2-binary
django-decouple
flower
